etc&refactor: customHooks 디렉토리 생성 및 Counter 컴포넌트 위치 수정
